+++ Source/Krypton Toolkit/Main/Krypton.Toolkit.Suite.Extended.Toast/Enumerations/Enumerations.cs
@@ -1,4 +1,12 @@
-﻿namespace Krypton.Toolkit.Suite.Extended.Toast
+﻿#region BSD License
+/*
+ * Use of this source code is governed by a BSD-style
+ * license or other governing licenses that can be found in the LICENSE.md file or at
+ * https://raw.githubusercontent.com/Krypton-Suite/Extended-Toolkit/master/LICENSE
+ */
+#endregion
+
+namespace Krypton.Toolkit.Suite.Extended.Toast
 {
     public enum ActionType
     {
@@ -39,5 +47,16 @@
     {
         LEFTTORIGHT,
         RIGHTTOLEFT
+    }
+
+    public enum InputBoxSystemSounds
+    {
+        ASTERISK = 0,
+        BEEP = 1,
+        EXCLAMATION = 2,
+        HAND = 3,
+        QUESTION = 4,
+        CUSTOM = 5,
+        NONE = 6
     }
 }